update made on enquiry form and create seprate files for styles
