+++ package.json
@@ -15,13 +15,15 @@
     "@mui/icons-material": "^5.11.16",
     "@mui/material": "^5.12.3",
     "axios": "^1.4.0",
+    "formik": "^2.4.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-images-viewer": "^1.7.1",
     "react-query": "^3.39.3",
     "react-router-dom": "^6.11.1",
     "react-simple-image-viewer": "^1.2.2",
-    "swiper": "^9.3.0"
+    "swiper": "^9.3.0",
+    "yup": "^1.2.0"
   },
   "devDependencies": {
     "@types/react": "^18.0.28",
