+++ src/components/banner/banner.jsx
@@ -8,97 +8,17 @@ import { ButtonDropdown } from "../common/dropdowns/dropdowns";
 //icons
 import HeadPhoneIcon from "../../assets/icons/HeadPhoneIcon.svg";
 import { Icon } from "../common/icon/icon";
-
-const BannerBox = styled(Box)(({ theme, image }) => ({
-  backgroundColor: theme.palette.primary.main,
-  width: "auto",
-  height: 416,
-  backgroundImage: `url(${image})`,
-  backgroundSize: "cover",
-  backgroundPosition: "center",
-  backgroundRepeat: "no-repeat",
-  display: "flex",
-  flexDirection: "column",
-  [theme.breakpoints.down("sm")]: {
-    height: 243,
-  },
-  justifyContent: "center",
-  position: "relative",
-}));
-
-const BannerMainText = styled(PrimaryText)(({ theme }) => ({
-  [theme.breakpoints.down("sm")]: {
-    fontSize: 22,
-  },
-}));
-
-const BannerSubText = styled(PrimarySubText)(({ theme }) => ({
-  [theme.breakpoints.down("sm")]: {
-    fontSize: 12,
-    padding: "0 1rem 0 1rem",
-  },
-}));
-
-const BannerFooter = styled(Box)(({ theme }) => ({
-  display: "flex",
-  position: "absolute",
-  bottom: "10%",
-  width: "93%",
-  justifyContent: "space-between",
-  marginLeft: "40px",
-  [theme.breakpoints.down("md")]: {
-    width: "90%",
-    marginLeft: "30px",
-  },
-  [theme.breakpoints.down("sm")]: {
-    width: "98%",
-    marginLeft: "5px",
-  },
-  [theme.breakpoints.down("sm")]: {
-    width: "94%",
-    marginLeft: "15px",
-  },
-}));
-
-const SubProgramBox = styled(Box)(() => ({
-  display: "flex",
-  flexDirection: "row",
-  gap: 10,
-  alignItems: "center",
-}));
-
-const SubProgramButton = styled(PrimaryButton)(() => ({}));
-
-const RecommendationButton = styled(SecondaryButton)(({ theme }) => ({
-  color: theme.palette.secondary.main,
-  [theme.breakpoints.down("sm")]: {
-    width: 142,
-    fontSize: 14,
-  },
-  [theme.breakpoints.down("xs")]: {
-    width: 142,
-    fontSize: 12,
-  },
-}));
-
-const TalkToExpertButton = styled(SecondaryButton)(({ theme }) => ({
-  ".MuiButton-startIcon": {
-    margin: 0,
-  },
-  [theme.breakpoints.down("sm")]: {
-    minWidth: 0,
-    width: 42,
-  },
-}));
-
-const RecommendationBox = styled(SecondaryButton)(({ theme }) => ({
-  display: "flex",
-  flexDirection: "row",
-  gap: 16,
-  [theme.breakpoints.down("sm")]: {
-    gap: 10,
-  },
-}));
+import {
+  BannerBox,
+  BannerMainText,
+  BannerSubText,
+  BannerFooter,
+  SubProgramBox,
+  SubProgramButton,
+  RecommendationButton,
+  TalkToExpertButton,
+  RecommendationBox,
+} from "./banner.styles";
 
 const subPrograms = ["Upper Back", "Lower Back", "Lower Back", "Lower Back", "Lower Back"];
 
@@ -125,9 +45,9 @@ export const Banner = ({ bannerData }) => {
           ) : (
             <>
               <SubProgramBox>
-                {subPrograms?.map((subProgram) => {
+                {subPrograms?.map((subProgram, index) => {
                   return (
-                    <SubProgramButton key={subProgram} variant="contained" color="buttonSecondary">
+                    <SubProgramButton key={index} variant="contained" color="buttonSecondary">
                       {subProgram}
                     </SubProgramButton>
                   );
