+++ src/components/emailValidation/emailValidation.jsx
@@ -1,63 +1,12 @@
 import { DialogContent, Dialog, Button, styled, TextField, Box } from "@mui/material";
 import { useState } from "react";
-import { PrimaryButton, SecondaryButton } from "../../components/common/buttons/button";
-import { PrimarySubText } from "../../components/common/typographies/typographies";
+import { PrimaryButton, SecondaryButton } from "../common/buttons/button";
+import { PrimarySubText } from "../common/typographies/typographies";
 import CloseIcon from "../../assets/icons/Close.svg";
 import ArrowIcon from "../../assets/icons/StyledArrow.svg";
-import { Icon } from "../../components/common/icon/icon";
+import { Icon } from "../common/icon/icon";
+import { HeaderBox, SubHeaderBox, ContentBox, StyleedDialogContent } from "./emailValidation.styles";
 
-const HeaderBox = styled(Box)(({ theme }) => ({
-  display: "flex",
-  flexDirection: "row",
-  justifyContent: "space-between",
-}));
-const SubHeaderBox = styled(Box)(({ theme }) => ({
-  display: "flex",
-  justifyContent: "unset",
-  gap: 5,
-}));
-const ContentBox = styled(Box)(({ theme }) => ({
-  display: "flex",
-  gap: 10,
-}));
-const StyledButton = styled(SecondaryButton)(({ theme }) => ({
-  height: 55,
-  lineHeight: 0,
-  fontSize: 16,
-  fontWeight: 700,
-  paddingTop: 12,
-  paddingBottom: 12,
-  paddingLeft: 22,
-  paddingRight: 22,
-  [theme.breakpoints.down("sm")]: {
-    width: 142,
-  },
-  [theme.breakpoints.down("xs")]: {
-    width: 142,
-    fontSize: 14,
-  },
-}));
-const StyleedDialogContent = styled(DialogContent)(({ theme }) => ({
-  background: "linear-gradient(180deg, rgba(105, 198, 172, 0.82) 0%, rgba(105, 198, 172, 0.5002) 100%)",
-  display: "flex",
-  flexDirection: "column",
-  overflow: "hidden",
-  height: "auto",
-}));
-const StyledTextField = styled(TextField)(({ theme }) => ({
-  borderRadius: 4,
-  input: {
-    background: theme.palette.primary.contrastText,
-    borderRadius: 4,
-  },
-}));
-const SubTitle = styled(PrimarySubText)(({ theme }) => ({
-  fontSize: 15,
-  fontWeight: 700,
-  color: theme.palette.primary.contrastText,
-  textAlign: "left",
-  paddingBottom: 18,
-}));
 export const SubContainerBox = () => {
   return (
     <SubHeaderBox>
